packages/ai/pyproject.toml
[project]
name = "coomb-ai"
version = "1.0.0"
description = "AI Platform for career optimization, resume analysis and professional mentorship"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "Yan Castro", email = "yan@coomb.io" }
]
keywords = ["ai", "resume", "career", "ats", "langchain", "crewai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    
    # AI/LLM
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "crewai>=0.80.0",
    "openai>=1.50.0",
    
    # Vector Store / RAG
    "qdrant-client>=1.12.0",
    "chromadb>=0.5.0",
    
    # PDF Generation
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    
    # Document Parsing
    "pdfplumber>=0.11.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    
    # Cache & Storage
    "redis>=5.0.0",
    "aiofiles>=24.0.0",
    "boto3>=1.35.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

